% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{catch_arcgislayer_error}
\alias{catch_arcgislayer_error}
\title{Retrieve status code from ArcGIS servers}
\usage{
catch_arcgislayer_error(furl)
}
\arguments{
\item{furl}{the base url.}
}
\value{
character message or invisible \code{NULL}
}
\description{
ArcGIS servers return status codes within the response body itself (as json).
\code{catch_arcgislayer_error} captures and returns status code message if there
is an error code. Otherwise it will return an invisible \code{NULL}. This is
primarily an internal function to help our data retrieval functions
fail gracefully in case of eorrs returns by the server.
}
\keyword{internal}
