% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nass.R
\name{check_census_year}
\alias{check_census_year}
\title{Check year argument}
\usage{
check_census_year(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{character}

\item{arg}{defaults to \code{rlang::caller_arg()}}

\item{call}{defaults to \code{rlang::caller_env()}}
}
\value{
error or nothing
}
\description{
Check year argument
}
\keyword{internal}
