% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_connectivity}
\alias{check_connectivity}
\title{Check internet connectivity}
\usage{
check_connectivity(host, call = rlang::caller_env())
}
\arguments{
\item{host}{url}

\item{call}{defaults to \code{\link[rlang:stack]{rlang::caller_env()}} and passed to \code{\link[cli:cli_abort]{cli::cli_inform}}}
}
\value{
\code{TRUE} or message with invisible \code{NULL}
}
\description{
Checks dns via curl and return cli message or invisible null.
}
\keyword{internal}
