% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_terra_gdal_config}
\alias{check_terra_gdal_config}
\title{Check terra gdalconfig AWS signin option}
\usage{
check_terra_gdal_config(call = rlang::caller_env())
}
\arguments{
\item{call}{environments that define where check_terra_gdal_config was called.}
}
\value{
called for side effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
No longer used internally by SELECTRdata.
Checks the config options for terra gdal AWS_NO_SIGN_REQUEST != "YES
}
\keyword{internal}
