% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_counties.R
\name{download_counties}
\alias{download_counties}
\title{Download TIGER county spatial data.}
\usage{
download_counties(template, output = tempfile(fileext = ".gpkg"))
}
\arguments{
\item{template}{A SpatRaster or SpatVector object. The extent of the returned object will match \code{template}.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}
}
\value{
A terra SpatVector object. If API resources are not available an invisible \code{NULL} is returned.
}
\description{
Shortcut function that downloads and extracts TIGER U.S. County boundaries
and returns them as a terra SpatVector object.
}
\examples{
# example code
\donttest{
## This example requires an internet connection to run
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)

counties <- download_counties(template = dem)
}

}
