% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nass.R
\name{download_nass_census}
\alias{download_nass_census}
\title{Wrapper for rnassqs}
\usage{
download_nass_census(
  group_desc = "LIVESTOCK",
  commodity_desc = "CATTLE",
  statisticcat_desc = "INVENTORY",
  short_desc = "CATTLE, INCL CALVES - INVENTORY",
  domain_desc = "TOTAL",
  state_alpha = "TX",
  county_name = "BRAZOS",
  year = "2022",
  ...
)
}
\arguments{
\item{group_desc}{character. See \code{rnassqs::parameter_values("group_desc")}.}

\item{commodity_desc}{character. See \code{rnassqs::parameter_values("commodity_desc")}.}

\item{statisticcat_desc}{character. See \code{rnassqs::parameter_values("statisticcat_desc")}.}

\item{short_desc}{character. See \code{rnassqs::parameter_values("short_desc")}.}

\item{domain_desc}{character. See \code{rnassqs::parameter_values("domain_desc")}.}

\item{state_alpha}{character. See \code{rnassqs::parameter_values("state_alpha")}.}

\item{county_name}{character. See \code{rnassqs::parameter_values("county_name")}.}

\item{year}{character. See \code{rnassqs::parameter_values("year")}. Expects one of: \code{c('2022', '2017', '2012', '2007', '2002', '1997')}.}

\item{...}{additional arguments carried to \code{rnassqs::nassqs()}.}
}
\value{
dataframe
}
\description{
Shortcut wrapper with arguments for commonly used parameters in rnassqs. Intended for internal use.
}
\keyword{internal}
