% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nass.R
\name{download_nass_livestock}
\alias{download_nass_livestock}
\title{Download county level total livestock}
\usage{
download_nass_livestock(
  state_alpha = NULL,
  county_name = NULL,
  year = "2022",
  ...
)
}
\arguments{
\item{state_alpha}{Two letter character abbreviation for state.}

\item{county_name}{Character, county name.}

\item{year}{A character to filter the NASS Agriculture Census year. Expects one of: \code{c('2022', '2017', '2012', '2007', '2002', '1997')}.}

\item{...}{Additional arguments passed to \code{rnassqs::nassqs()}}
}
\value{
A dataframe.
}
\description{
Provides a wrapper to \code{rnassqs::nassqs()} with simplified arguments
for making queries that return county level livestock estimates. This
function requires an API key from USDA NASS Quickstats. See details for
more information.
}
\details{
This function requires an API key to make requests on the USDA NASS QuickStats service.
\enumerate{
\item Obtain an API key from \url{https://quickstats.nass.usda.gov/api/}.
\item Set the \code{NASSQS_TOKEN} variable in your global environment. This can be done with \verb{Sys.setenv('NASSQS_TOKEN' = <your api key>)}, or \code{usethis::edit_r_environment()}.
\item Restart your session.
}
}
\examples{
\dontshow{if (has_nass_token()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
download_nass_livestock(state_alpha = "TX", county_name = "Brazos", year = "2022")
\dontshow{\}) # examplesIf}
}
