% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_internet_2}
\alias{has_internet_2}
\title{nslookup wrapper}
\usage{
has_internet_2(host)
}
\arguments{
\item{host}{hostname}
}
\value{
logical value
}
\description{
Wrapper for \code{\link[curl:nslookup]{curl::nslookup}} that returns logical without error.
}
\keyword{internal}
