% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_nass_token}
\alias{has_nass_token}
\title{Is there a NASS token?}
\usage{
has_nass_token()
}
\value{
logical
}
\description{
Checks if a NASS token has been set in the user environment.
}
\examples{
has_nass_token()
}
