% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{thompson}
\alias{thompson}
\title{Thompsons Creek Elevation Data}
\description{
Elevation raster of the Thompsons Creek watershed outside of
College Station, Texas. This data is obtained from the hydroreinforced
digital elevation model (DEM) rasters that are part of the United States
Geological Survey (USGS) and United States Environmental Protection Agency
(EPA) National Hydrography Dataset Plus (NHDPlus).
}
\examples{
\donttest{
thompson <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
terra::rast(thompson)
}
}
\seealso{
\url{https://nhdplus.com/NHDPlus/NHDPlusV2_home.php}
}
