% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census_blocks.R
\name{download_census_blocks}
\alias{download_census_blocks}
\title{Download U.S. Census block boundaries}
\usage{
download_census_blocks(
  template,
  year = "2020",
  page_size = NULL,
  output = tempfile(fileext = ".gpkg")
)
}
\arguments{
\item{template}{A SpatRaster or SpatVector object.}

\item{year}{A character value. Any of the following values should work: \code{c("2000","2010","2020")}.}

\item{page_size}{a numeric value passed to \code{arcgislayers::arcselect()}. Defaults to NULL. Useful when the requests returns a 500 error code.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}
}
\value{
A terra SpatVector object. If API resources are not available an invisible \code{NULL} is returned.
}
\description{
Downloads U.S. Census block boundaries and associated information for
blocks within the spatial extent of the template SpatRaster.
}
\examples{
# example code
\donttest{
## This example requires an internet connection to run
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)

blocks <- download_census_blocks(template = dem)
blocks
}

}
