% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census_blocks.R
\name{download_census_blocks}
\alias{download_census_blocks}
\title{Download U.S. Census block boundaries}
\usage{
download_census_blocks(
  template,
  year = "2020",
  output = tempfile(fileext = ".gpkg")
)
}
\arguments{
\item{template}{A SpatRaster object.}

\item{year}{A character value. Any of the following values should work: \code{c("2000","2010","2020")}.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}
}
\value{
A terra SpatVector object.
}
\description{
Downloads U.S. Census block boundaries and associated information for
blocks within the spatial extent of the template SpatRaster.
}
