% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_counties.R
\name{download_counties}
\alias{download_counties}
\title{Download TIGER county spatial data.}
\usage{
download_counties(template, output = tempfile(fileext = ".gpkg"))
}
\arguments{
\item{template}{A SpatRaster object.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}
}
\value{
A terra SpatVector object.
}
\description{
Shortcut function that downloads and extracts TIGER U.S. County boundaries
and returns them as a terra SpatVector object.
}
