% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nlcd.R
\name{download_nlcd}
\alias{download_nlcd}
\title{Download and write annual NLCD raster data to file.}
\usage{
download_nlcd(
  template,
  year = "2021",
  dataset = "LndCov",
  landmass = "CU",
  output = tempfile(fileext = ".tiff"),
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{template}{A SpatRaster object defining the spatial extent of the returned NLCD raster.}

\item{year}{character, expects a value between \code{1986:2023}.}

\item{dataset}{Character. Expects \code{c("LndCov","LndChg","LndCnf","FctImp","ImpDsc","SpcChg")}.}

\item{landmass}{Character, one of: \code{c("CU", "AK", "HI")}.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{overwrite}{logical. If \code{TRUE}, filename is overwritten}

\item{verbose}{Logical, if \code{TRUE} informative messages will be printed.}

\item{...}{additional arguments for for writing files, see \code{terra::writeRaster()}}
}
\value{
A SpatRaster object with file written to \code{output}
}
\description{
Downloads and writes an NLCD SpatRaster to file with extents defined by \code{template}. This function downloads the annualized NLCD data products. See \url{https://www.mrlc.gov/data/project/annual-nlcd} for more information.
}
