% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_NPDES_permits.R
\name{download_NPDES_permits}
\alias{download_NPDES_permits}
\title{Download NPDES permits}
\usage{
download_NPDES_permits(
  template,
  permit_component = "POT",
  permit_status = "EFF",
  output = tempfile(fileext = ".gpkg")
)
}
\arguments{
\item{template}{A SpatRaster object.}

\item{permit_component}{A character vector with one or more of the following: \code{PRE} (pretreatment), \code{CAF} (CAFO), \code{CSO} (CSO), \code{POT} (Publicly Owned Treatment Works, the default), \code{BIO} (Biosolids), \code{SWS} (Stormwater Small MS4), \code{SWM} (Stormwater Medium/Large MS4), \code{SWI} (Stormwater Industrial), \code{SWC} (Stormwater Construction).}

\item{permit_status}{A character vector with one or more of the following: \code{EFF} (effective, the default), \code{EXP} (expired), \code{PND} (pending), \code{TRM} (terminated), \code{RET} (retired), \code{NON} (not needed), \code{ADC} (administratively continued).}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}
}
\value{
SpatVector object
}
\description{
Downloads NPDES permits from EPA ECHO API services within the bounds of the
SpatRast template.
}
\examples{
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)
download_NPDES_permits(dem)

}
