% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_urban_areas.R
\name{download_urban_areas}
\alias{download_urban_areas}
\title{Download U.S. Urban Areas}
\usage{
download_urban_areas(
  template,
  page_size = NULL,
  output = tempfile(fileext = ".gpkg")
)
}
\arguments{
\item{template}{A SpatRaster object.}

\item{page_size}{a numeric value passed to \code{arcgislayers::arcselect()}. Defaults to NULL. Useful when the requests returns a 500 error code.}

\item{output}{A character file path specifying where the \code{SpatVector} output object should be written. Defaults to a temporary file.}
\value{
A terra SpatVector object. If API resources are not available an invisible \code{NULL} is returned.
}
\description{
Downloads a U.S. Census designated urban areas intersecting the spatial extent of the template SpatRaster.
}
\examples{
# example code
\donttest{
## This example requires an internet connection to run
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)

ua <- download_urban_areas(template = dem)
ua
}

}
