% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gdal_config.R
\name{set_gdal_config}
\alias{set_gdal_config}
\title{Set GDAL configuration options}
\usage{
set_gdal_config(option, value)
}
\arguments{
\item{option}{Character name of a configuration option.}

\item{value}{Character value to set for the option. \code{value = ""} (empty string) will unset previously set values.}
}
\value{
No return value, called for side effects.
}
\description{
A convience functions for simultaneously setting the GDAL runtime
configuration options in both terra and gdalraster.
}
\examples{
set_gdal_config("GDAL_CACHEMAX", "10\%")
## unset
set_gdal_config("GDAL_CACHEMAX", "")
}
\seealso{
\code{\link[terra:gdal]{terra::setGDALconfig()}} and \code{\link[gdalraster:set_config_option]{gdalraster::set_config_option()}}
}
