% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wbd}
\alias{wbd}
\title{Thompsons Creek Watershed Boundary}
\description{
Polygon boundary of the Thompsons Creek watershed (Collecge Station, Texas).
This watershed boundary was created using the hydroreinformced raster from
the United States Geological Survey (USGS) and United States Environmental
Protection Agency (EPA) National Hydrography Dataset Plus (NHDPlus) dataset.
}
\examples{
\donttest{
gpkg <- system.file("extdata", "thompsoncreek.gpkg", package = "SELECTRdata")
wbd <- terra::vect(gpkg, layer = "wbd")
}
}
\seealso{
\code{\link[=thompson]{thompson()}}
}
